requirements.txt
requests>=2.31.0
PyJWT>=2.8.0
shodan>=1.30.0
termcolor>=2.3.0
boto3>=1.34.0
botocore>=1.34.0
aiohttp>=3.9.0
aiofiles>=23.2.0
dnspython>=2.4.2
pyOpenSSL>=23.3.0
certifi>=2023.11.17
